simnode: generate correct build version for Nomad client attrs.
The Nomad version attr is used when making RPC calls from the
server to the client. Not correctly populating this meant that
these RPC calls could not be made.

The change also moves the build info generation into its own
package, so it can run only once per trigger on the simnode app.