config: add config file parsing and parameters for client cfg.
This change adds the ability to load configuration using a config
file, written in HCL or JSON. The new file method also includes
additional config parameters to control the simulated Nomad
clients.

I changed the base flags to be a little more descriptive and align
with the config file entry.