allocrunner: ensure alloc state updates are handled correctly. (#22)
Implementing the GetUpdatePriority function is not enough to
ensure updates are handling in a similar fashion to real clients.

The alloc runner needs to be careful in how and when it updates
its own state. Specifically, we do not want to directly force
the client status information and instead have this populated once
updates have been sent to the servers. The new functionality helps
ensure updates are sent at the correct pacing, and handled in a
manner more similar to real clients.