lock allocstate before adding task event
We don't lock the alloc state before adding a task event. This can result in a
panic when we emit allocation metrics if it happens concurrently, because the
length of the slice isn't atomically updated with the contents so we end up
indexing an event that's been added before the slice is valid.